ODIN-197: tenant funding source verification (#206)

+++ src/tenant/fundingSource/logic.ts
@@ -3,18 +3,24 @@ import {CreateTenantFundingSourceRequest} from './models';
 import * as dwolla from '../../dwolla';
 import {TenantService} from '../service';
 import {Tenant} from '../models';
+import {Logic} from '../../logic';
+import {Errors} from 'typescript-rest';
+import {VerificationStatuse} from '../../foundingSource/models';
 
 
 @AutoWired
-export class CreateTenantFundingSourceLogic {
+export class CreateTenantFundingSourceLogic extends Logic {
     @Inject private dwollaClient: dwolla.Client;
     @Inject private tenantService: TenantService;
 
     async execute(request: CreateTenantFundingSourceRequest, tenantId: string): Promise<Tenant> {
         const tenant = await this.tenantService.get(tenantId);
+        if (!tenant) {
+            throw new Errors.NotFoundError('Tenant not found');
+        }
 
         if (tenant.fundingSourceUri) {
-            throw new Error('Could not add many funding sources.');
+            throw new Errors.NotAcceptableError('Could not add more funding sources');
         }
 
         try {
@@ -37,21 +43,33 @@ export class CreateTenantFundingSourceLogic {
 }
 
 @AutoWired
-export class GetTenantFundingSourceLogic {
-    @Inject private service: TenantService;
+export class GetTenantFundingSourceLogic extends Logic {
+    @Inject private tenantService: TenantService;
 
     async execute(tenantId: string): Promise<Tenant> {
-        return await this.service.get(tenantId);
+        const tenant = await this.tenantService.get(tenantId);
+        if (!tenant) {
+            throw new Errors.NotFoundError('Tenant not found');
+        }
+
+        return tenant;
     }
 }
 
 @AutoWired
-export class DeleteTenantFundingSourcesLogic {
-    @Inject private service: TenantService;
+export class DeleteTenantFundingSourcesLogic extends Logic {
+    @Inject private tenantService: TenantService;
     @Inject private client: dwolla.Client;
 
     async execute(tenantId: string) {
-        const tenant = await this.service.get(tenantId);
+        const tenant = await this.tenantService.get(tenantId);
+        if (!tenant) {
+            throw new Errors.NotFoundError('Tenant not found');
+        }
+
+        if (!tenant.fundingSourceUri) {
+            throw new Errors.NotFoundError('Funding source not found');
+        }
 
         await this.client.deleteFundingSource(tenant.fundingSourceUri);
 
@@ -60,6 +78,62 @@ export class DeleteTenantFundingSourcesLogic {
         tenant.fundingSourceName = null;
         tenant.fundingSourceRouting = null;
 
-        await this.service.update(tenant);
+        await this.tenantService.update(tenant);
+    }
+}
+
+@AutoWired
+export class InitiateTenantFundingSourceVerificationLogic extends Logic {
+    @Inject private tenantService: TenantService;
+    @Inject private client: dwolla.Client;
+
+    async execute(tenantId: string) {
+        const tenant = await this.tenantService.get(tenantId);
+        if (!tenant.fundingSourceUri) {
+            throw new Errors.NotFoundError('Funding source not found');
+        }
+
+        if (tenant.fundingSourceVerificationStatus) {
+            throw new Errors.NotAcceptableError('Funding source verification cannot be initiated');
+        }
+
+        if (!await this.client.createFundingSourceMicroDeposit(tenant.fundingSourceUri)) {
+            throw new Errors.NotAcceptableError('Funding source verification initiation failed');
+        }
+
+        tenant.fundingSourceVerificationStatus = VerificationStatuse.initiated;
+        await this.tenantService.update(tenant);
+    }
+}
+
+@AutoWired
+export class VerifyTenantFundingSourceLogic extends Logic {
+    @Inject private tenantService: TenantService;
+    @Inject private client: dwolla.Client;
+
+    async execute(amount1, amount2: number, tenantId: string) {
+        const tenant = await this.tenantService.get(tenantId);
+        if (!tenant.fundingSourceUri) {
+            throw new Errors.NotFoundError('Funding source not found');
+        }
+
+        if (tenant.fundingSourceVerificationStatus != VerificationStatuse.initiated) {
+            throw new Errors.NotAcceptableError('Funding source verification not initiated');
+        }
+
+        try {
+            await this.client.verifyFundingSourceMicroDeposit(tenant.fundingSourceUri, amount1, amount2);
+        } catch (e) {
+            if (e instanceof dwolla.DwollaRequestError) {
+                // I HATE DWOLLA SO MUCH SINCE THEY MADE ME DO IT!
+                e.message = e.message.replace(/\/value/g, '');
+                throw e.toValidationError();
+            }
+
+            throw e;
+        }
+
+        tenant.fundingSourceVerificationStatus = VerificationStatuse.completed;
+        await this.tenantService.update(tenant);
     }
 }
